Integrated with Servless.com platform. Added additional logging to index.js Updated readm Other duties as required.
