--- app.local.js
@@ -1,5 +1,0 @@
-const app = require('./app')
-const port = 3000
-
-app.listen(port)
-console.log(`listening on http://localhost:${port}`)
