+++ package.json
@@ -5,42 +5,13 @@
   "repository": "https://github.com/lukeburns/netlify-cms-github-lambda",
   "license": "MIT",
   "dependencies": {
-    "aws-serverless-express": "^3.3.5",
-    "dotenv": "^2.0.0",
-    "express": "^4.15.2",
+    "dotenv": "^6.2.0",
+    "express": "^4.16.4",
     "randomstring": "^1.1.5",
-    "simple-oauth2": "^1.2.0"
+    "serverless-http": "^1.8.0",
+    "simple-oauth2": "^1.6.0"
   },
-  "config": {
-    "s3BucketName": "YOUR_UNIQUE_BUCKET_NAME",
-    "region": "YOUR_AWS_REGION",
-    "cloudFormationStackName": "AwsServerlessExpressStack",
-    "functionName": "YOUR_SERVERLESS_EXPRESS_LAMBDA_FUNCTION_NAME",
-    "accountId": "YOUR_ACCOUNT_ID"
-  },
-  "scripts": {
-    "start": "node app.local.js",
-    "config": "node ./scripts/configure.js",
-    "deconfig": "node ./scripts/deconfigure.js",
-    "local": "node scripts/local",
-    "invoke-lambda": "aws lambda invoke --function-name $npm_package_config_functionName --region $npm_package_config_region --payload file://api-gateway-event.json lambda-invoke-response.json && cat lambda-invoke-response.json",
-    "create-bucket": "aws s3 mb s3://$npm_package_config_s3BucketName --region $npm_package_config_region",
-    "delete-bucket": "aws s3 rb s3://$npm_package_config_s3BucketName --region $npm_package_config_region",
-    "package": "aws cloudformation package --template ./cloudformation.yaml --s3-bucket $npm_package_config_s3BucketName --output-template packaged-sam.yaml --region $npm_package_config_region",
-    "deploy": "aws cloudformation deploy --template-file packaged-sam.yaml --stack-name $npm_package_config_cloudFormationStackName --capabilities CAPABILITY_IAM --region $npm_package_config_region",
-    "package-deploy": "npm run package && npm run deploy",
-    "delete-stack": "aws cloudformation delete-stack --stack-name $npm_package_config_cloudFormationStackName --region $npm_package_config_region",
-    "setup": "npm install && (aws s3api get-bucket-location --bucket $npm_package_config_s3BucketName --region $npm_package_config_region || npm run create-bucket) && npm run package-deploy",
-    "win-config": "npm run config",
-    "win-deconfig": "npm run deconfig",
-    "win-local": "npm run local",
-    "win-invoke-lambda": "aws lambda invoke --function-name %npm_package_config_functionName% --region %npm_package_config_region% --payload file://api-gateway-event.json lambda-invoke-response.json && cat lambda-invoke-response.json",
-    "win-create-bucket": "aws s3 mb s3://%npm_package_config_s3BucketName% --region %npm_package_config_region%",
-    "win-delete-bucket": "aws s3 rb s3://%npm_package_config_s3BucketName% --region %npm_package_config_region%",
-    "win-package": "aws cloudformation package --template ./cloudformation.yaml --s3-bucket %npm_package_config_s3BucketName% --output-template packaged-sam.yaml --region %npm_package_config_region%",
-    "win-deploy": "aws cloudformation deploy --template-file packaged-sam.yaml --stack-name %npm_package_config_cloudFormationStackName% --capabilities CAPABILITY_IAM --region %npm_package_config_region%",
-    "win-package-deploy": "npm run win-package && npm run win-deploy",
-    "win-delete-stack": "aws cloudformation delete-stack --stack-name %npm_package_config_cloudFormationStackName% --region %npm_package_config_region%",
-    "win-setup": "npm install && (aws s3api get-bucket-location --bucket %npm_package_config_s3BucketName% --region %npm_package_config_region% || npm run win-create-bucket) && npm run win-package-deploy"
+  "devDependencies": {
+    "serverless-domain-manager": "^2.6.10"
   }
 }
