+++ serverless.yml
@@ -1,0 +1,35 @@
+# serverless.yml
+
+service: netlify-cms-auth
+
+provider:
+  name: aws
+  runtime: nodejs8.10
+  stage: prod
+  region: us-east-1
+
+# Creating single function, app which uses exported handler from index.js
+# Has a few groovy http triggers
+functions:
+  app:
+    handler: index.handler
+    environment:
+      NODE_ENV: production
+      OAUTH_CLIENT_ID: OAUTH_CLIENT_ID_FROM_GITHUB #Get from https://github.com/settings/developers
+      OAUTH_CLIENT_SECRET: OAUTH_CLIENT_SECRET_FROM_GITHUB #Get from hhttps://github.com/settings/developers
+      REDIRECT_URL: https://<auth.yourcompany.com>/callback
+      AUTH_TARGET: _blank
+    events:
+      - http: ANY /
+      - http: 'ANY {proxy+}'
+
+plugins:
+  - serverless-domain-manager
+custom:
+  customDomain:
+    domainName: <auth.yourcompany.com>
+    basePath: 'prod'
+    stage: ${self:provider.stage}
+    createRoute53Record: false
+    # Set createRoute53Record to true if you are using route 53 to manage your domain's DNS
+
